fix(hooks): preserve Windows shell contracts
Problem: the first PR revision dropped auto-detected Bash from assembled plugin hooks, converted every implicit shell path to POSIX form, and exceeded the repository lint budgets.

Root cause: plugin assembly copied the manifest shell instead of the normalized execution contract, while path conversion ran before the actual shell/script decision. New logic and tests were also appended to already over-budget files.

Fix: propagate the normalized shell, convert paths only for explicit or shebang-confirmed Bash scripts, preserve native Windows implicit hook paths, and move the POSIX hook implementation and tests into dedicated files.

Verification: go test ./internal/...; go test -race ./internal/hook; GOOS=windows GOARCH=amd64 go test -c ./internal/hook -o /tmp/reasonix-hook.test.exe; go run ./tools/repolint; git diff --check.
